Fix integer division in floating point context

+++ src/me/midest/logic/coupling/VisitsValue.java
@@ -18,8 +18,8 @@ public class VisitsValue {
             return 0;
         int vSize = visits.size();
         int vPairsSize = vSize * ( vSize - 1 ) / 2;
-        double L = allPairsSum( visits )/ vPairsSize;
-        double R = rulesSum( visits ) / vSize;
+        double L = allPairsSum( visits )/ (double) vPairsSize;
+        double R = rulesSum( visits ) / (double) vSize;
         return L + ( DELTA * R / vSize );
     }
